requirements.txt
pymongo==3.12.0
colorama==0.4.4
pyfiglet==0.8.post1
pycryptodome==3.11.0